forecast_tournament/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.0.0
gspread>=5.0.0
google-auth>=2.0.0
